clened up button structure

+++ src/ui/rectangle/Rectangle.h
@@ -21,7 +21,7 @@ public:
   /// @brief Sets the position of the rectangle
   /// @param x The new x
   /// @param y The new y
-  void SetPosition(int x, int y);
+  void SetPosition(float x, float y);
 
 protected: // Bro t'he cambiat això. Ho necessito :/
   void Render() override;
@@ -29,6 +29,6 @@ protected: // Bro t'he cambiat això. Ho necessito :/
   /// @param x The x coord
   /// @param y The y coord
   /// @return Whether it is inside the rectangle
-  bool IsWithinRect(int x, int y);
+  bool IsWithinRect(float x, float y);
 };
 } // namespace Chess::Rendering::Elements
